vc/amd/sb800: Cast to UINT32 for shift out of bounds fix
It’s defined as `unsigned char`.

    SB800: sb_Before_Pci_Init
    shift out of bounds src/vendorcode/amd/cimx/sb800/SBCMN.c:643:53
    ubsan: unrecoverable error.

Found-by: UBSAN
Reviewed-on: https://review.coreboot.org/c/coreboot/+/51284
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/vendorcode/amd/cimx/sb800/SBCMN.c
@@ -640,12 +640,12 @@ abcfgTbl (
 
   while ( (pABTbl->regType) != 0xFF ) {
     if ( pABTbl->regType > AXINDC ) {
-      ddValue = pABTbl->regIndex | (pABTbl->regType << 29);
+      ddValue = pABTbl->regIndex | ((UINT32) pABTbl->regType << 29);
       writeAlink (ddValue, ((readAlink (ddValue)) & (0xFFFFFFFF^ (pABTbl->regMask))) | pABTbl->regData);
     } else {
-      ddValue = 0x30 | (pABTbl->regType << 29);
+      ddValue = 0x30 | ((UINT32) pABTbl->regType << 29);
       writeAlink (ddValue, pABTbl->regIndex);
-      ddValue = 0x34 | (pABTbl->regType << 29);
+      ddValue = 0x34 | ((UINT32) pABTbl->regType << 29);
       writeAlink (ddValue, ((readAlink (ddValue)) & (0xFFFFFFFF^ (pABTbl->regMask))) | pABTbl->regData);
     }
     ++pABTbl;
